Update README.md
a single terraform binary file is in ./terraform directory. It's not able to create a directory with the same name of the binary file name. Or the command you ran is not in the terraform but outside of the terraform directory. I stuck here logs as follow:

$ terraform apply
Acquiring state lock. This may take a few moments...

Error: Variables not allowed

  on terraform.tfvars line 1:
   1: project = contacttracing-275122

Variables may not be used here.